concierge/lambda_src/requirements.txt
# Requirements for Lambda functions
google-cloud-firestore
google-generativeai
flask-socketio[client]
websockets
firebase-admin
twilio
livekit